Update the Mac instructions to include an install_name_tool step
This changes the Mach-O library ID to match the install path. The
library ID is copied by the linker when Subsurface is linked against
Marble and can be seen with otool -L. That's also what macdeployqt
uses to find the dependencies: if the file matching the name in otool
-L exists, macdeployqt will deploy the library and update the search
paths.

Conclusion: after this patch, the "make mac-deploy" command will also
deploy Marble correctly.

Signed-off-by: Thiago Macieira <thiago@macieira.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>