Ruler: display maximum descent and ascent speed
While playing with the ruler, I figured that it happened only once that
my minimum speed was not 0 for a segment long enough to get any display.
And it was a "dive" in an hyperbaric chamber...

This patch replaces the minimum speed with the maximum descent speed and
the maximum speed with the maximum ascent speed.

Signed-off-by: Alexandre Belloni <alexandre.belloni@piout.net>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>