Globe: improve the flag placement and size
We can position our old flag in a PNG with double the size of the
original flag - 48x48px, so that the pin is exactly at the center
of the PNG.

This way when the user clicks on a map location the pin should
be placed exactly there.

Patch also makes the flag bigger.

Fixes #335.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>