Include the user-manual.html file in the git tree
This way when building Subsurface asciidoc is not necessary at all - except if
you want to work on the user manual and want to see the result of your work.

Previously the build wouldn't fail without asciidoc, but there would be no
working user manual when pressing F1 from within the application. This solves
the problem by shipping the HTML file with the sources.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>