When merging dives, keep dive ids in mind
We try to do the most reasonable thing. If you have an existing dive and
merge a just downloaded dive with it - the you want the existing id.
If you merge two consecutive dives, then the start of the merged dive is
the earlier of those two dives so we keep that id.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>