Make the 'Gtk' branch compile on Fedora 20
This makes our old Gtk branch work with F20, using gtk3.

Not seriously tested, and there are a few warnings, but it seems to
superficially work. Since the gtk3 support infrastructure was already
merged long ago, there's only one actual change to the source code,
the rest is all just the package name changes for gtk3 and the change
to the osmgpsmaps naming.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

 Makefile  | 8 ++++----
 gtk-gui.c | 2 +-
 2 files changed, 5 insertions(+), 5 deletions(-)
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>